requirements.txt
google-api-python-client==2.114.0
google-auth==2.27.0
requests==2.31.0
beautifulsoup4==4.12.3
